requirements.txt
astroid==2.9.3
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.5.1
google-cloud-bigquery==2.34.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
identify==2.4.11
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.0
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.17.0
proto-plus==1.20.3
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
